# src/main/resources/static/swagger.yml
openapi: 3.0.4
info:
  title: itPlace API Document
  description: |-
    지도 기반 내 주변 LG유플러스 멤버십 혜택 정보 조회 서비스 'itPlace'의 API 명세서입니다.
    
    모든 API 응답은 표준 응답 포맷(ApiResponse)을 따릅니다.
  version: 1.0.0
servers:
  - url: https://user-api.itplace.click/api/v1
    description: 개발 서버

  - url: http://localhost:8080/api/v1
    description: 로컬 테스트 서버
...
paths:
...
  /favorites/benefits/{benefitId}:
    get:
      tags:
        - Favorites
      summary: 즐겨찾기 혜택 상세 조회
      description: 혜택 ID를 기반으로 상세 정보를 조회합니다.
      operationId: getBenefitDetail
      parameters:
        - name: benefitId
          in: path
          required: true
          description: 혜택 ID
          schema:
            type: integer
      responses:
        '200':
          description: 혜택 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: FAVORITE_BENEFIT_DETAIL_SUCCESS
                status: OK
                message: 혜택 상세 정보 조회 성공
                data:
                  benefitId: 3
                  benefitName: "CGV"
                  benefitDescription: "영화 그 이상의 감동 CGV"
                  benefitLimit: "VVIP/VIP 등급 전용\r VIP콕 내 무료예매 연3회/1+1예매 연9회(총 12회)\r (월 1회 사용 가능, CGV/메가박스 중 택 1)"
                  partnerName: "CGV"
                  partnerImage: "https://itplacepartners.s3.ap-northeast-2.amazonaws.com/img/cgv.png"
                  tiers:
                    - grade: VIP콕
                      isAll: false
                      context: "무료 예매 연 3회 / 1+1 예매 연 9회(총 12회)\r\n(월 1회 사용 가능, CGV/메가박스 중 택 1)\r\n※ 1+1이란? 1매 유료로 구매 시, 1매 무료 제공"
                      discountValue: 14000
                timestamp: '2025-07-15T12:00:00Z'

        '404':
          description: 혜택 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: BENEFIT_NOT_FOUND
                status: NOT_FOUND
                message: 존재하지 않는 혜택입니다.
                data: null
                timestamp: '2025-07-16T15:00:00.000Z'

  # =====================================================
  # User 관련
  # =====================================================
...
  /benefit:
    get:
      tags:
        - Benefit
      summary: 전체 혜택 페이지에서 혜택 조회
      description: 대분류, 카테고리, 정렬(인기순), 필터링, 검색, 페이징 조건에 따라 혜택을 조회합니다.
      operationId: getBenefits
      parameters:
        - name: mainCategory
          in: query
          description: VIP콕/기본혜택
          required: true
          schema:
            type: string
            enum: [ VIP_COCK, BASIC_BENEFIT ]

        - name: category
          in: query
          description: 카테고리 이름(엑티비티, 푸드, ...)
          required: false
          schema:
            type: string

        - name: filter
          in: query
          description: 온라인/오프라인 필터
          required: false
          schema:
            type: string
            enum: [ ONLINE, OFFLINE ]

        - name: sort
          in: query
          description: 인기순 정렬. 기본 정렬도 인기순이지만 필터 선택을 통한 안가순 정렬도 가능.
          required: false
          schema:
            type: string
            default: POPULARITY

        - name: keyword
          in: query
          description: 검색어(혜택이름에서 결과를 찾음)
          required: false
          schema:
            type: string

        - name: page
          in: query
          description: 페이지 번호 (0부터 시작)
          required: false
          schema:
            type: integer
            default: 0

        - name: size
          in: query
          description: 한 페이지당 결과 수. 기본값은 12
          required: false
          schema:
            type: integer
            default: 12

      responses:
        '200':
          description: 혜택 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedResponse_BenefitListResponse'
              example:
                data:
                  content:
                    - benefitId: 34
                      benefitName: "배스킨라빈스"
                      mainCategory: "BASIC_BENEFIT"
                      usageType: "OFFLINE"
                      partnerId: 30
                      category: "푸드"
                      image: ""
                      tierBenefits:
                        - grade: "BASIC"
                          context: "패밀리 사이즈(5가지 맛) 4천원 할인"
                          isAll: true
                        - grade: "VIP"
                          context: "패밀리 사이즈(5가지 맛) 4천원 할인"
                          isAll: true
                        - grade: "VVIP"
                          context: "패밀리 사이즈(5가지 맛) 4천원 할인"
                          isAll: true
                      isFavorite: true
                      favoriteCount: 1200
                    - benefitId: 30
                      benefitName: "GS25"
                      mainCategory: "BASIC_BENEFIT"
                      usageType: "OFFLINE"
                      partnerId: 112
                      category: "생활/편의"
                      image: ""
                      tierBenefits:
                        - grade: "BASIC"
                          context: "1천원 당 50원 할인 \r\n일 최대 1만원까지 할인 가능\r\n월 최대 10만원까지 할인 가능\r\n※제휴사 지정하는 행사상품은 할인제외"
                          isAll: false
                        - grade: "VIP"
                          context: "1천원 당 100원 할인 \r\n일 최대 2만원까지 할인 가능\r\n월 최대 10만원까지 할인 가능\r\n※제휴사 지정하는 행사상품은 할인제외"
                          isAll: false
                        - grade: "VVIP"
                          context: "1천원 당 100원 할인 \r\n일 최대 2만원까지 할인 가능\r\n월 최대 10만원까지 할인 가능\r\n※제휴사 지정하는 행사상품은 할인제외"
                          isAll: false
                      isFavorite: true
                      favoriteCount: 1000
                  currentPage: 0
                  totalPages: 11
                  totalElements: 121
                  hasNext: true
                code: BENEFIT_LIST_SUCCESS
                status: OK
                message: 혜택 목록 조회에 성공했습니다.
                timestamp: '2025-07-04T15:30:00.000Z'
  /benefit/{benefitId}:
    get:
      tags:
        - Benefit
      summary: 혜택 상세 조회
      description: 혜택 ID를 통해 혜택의 상세 정보를 조회합니다.
      operationId: getBenefitDetail
      parameters:
        - name: benefitId
          in: path
          required: true
          description: 조회할 혜택의 ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 혜택 상세 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BenefitDetailResponse'
              example:
                data:
                  benefitId: 26
                  benefitName: "딸기콩"
                  description: "딸기콩은 웅진씽크빅에서 만든 국내 최대 어린이 오디오북 플랫폼으로 7,000여 개의 어린이 전문 콘텐츠를 기반으로 아이의 연령과 관심 영역에 맞춘 큐레이션을 제공합니다."
                  benefitLimit: "VIP콕 내 제휴사 통합 월 1회, 연 12회 가능"
                  manual: "멤버십앱 VIP 특별혜택 > 혜택 > 딸기콩 > 하단 '바로가기' 클릭 > 딸기콩 연동 후 등록\n\n\n\n\n*꼭 확인하세요\n\n\n- 딸기콩 구독권 또는 쿠폰 보유 시, 혜택 적용이 불가합니다.\n\n- 딸기콩 구독권 또는 쿠폰을 보유하고 계실 경우 딸기콩 1:1 문의 또는 웅진 고객센터(1577-1500)를 통해 혜택 적용이 가능합니다.\n\n- 본 혜택은 U+멤버십 혜택이 매달 자동 적용되지 않으며, 고객님께서 해당 페이지를 통해 매달 직접 신청하셔야 합니다.\n\n- 만 19세 이상 고객만 가입할 수 있습니다. \n\n\n\n대표문의\n\n딸기콩 고객센터 1577-1500"
                  url: "https://www.ddalgicong.com/login/login.html"
                  partnerName: "딸기콩"
                  image: ""
                code: BENEFIT_DETAIL_SUCCESS
                status: OK
                message: 혜택 상세 조회에 성공했습니다.
                timestamp: '2025-07-16T12:00:00.000Z'

        '404':
          description: 혜택을 찾을 수 없습니다
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                data: null
                code: BENEFIT_NOT_FOUND
                status: NOT_FOUND
                message: 존재하지 않는 혜택입니다.
                timestamp: '2025-07-16T12:00:00.000Z'
  /benefit/map-detail:
    get:
      tags:
        - Benefit
      summary: 제휴처 지도에서 혜택 상세 조회
      description: 지점(storeId), 제휴사(partnerId), 혜택 대분류(mainCategory)를 기준으로 혜택 상세 정보를 조회합니다.
      operationId: getMapBenefitDetail
      parameters:
        - name: storeId
          in: query
          required: true
          description: 지점 ID
          schema:
            type: integer
            format: int64
        - name: partnerId
          in: query
          required: true
          description: 제휴사 ID
          schema:
            type: integer
            format: int64
        - name: mainCategory
          in: query
          required: true
          description: 혜택 대분류
          schema:
            type: string
            enum: [ VIP_COCK, BASIC_BENEFIT ]
      responses:
        '200':
          description: 혜택 상세 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MapBenefitDetailResponse'
              example:
                data:
                  benefitId: 21
                  benefitName: "롯데월드 어드벤처"
                  mainCategory: "VIP_COCK"
                  manual: "결제 시 직원에게 멤버십 카드 제시\n\n\n\n*꼭 확인하세요\n\n- 기타 우대혜택 및 카드할인과 중복적용되지 않습니다.\n\n- 현장 구매 티켓은 당일에 한하여 이용가능합니다. \n\n- 본인 소유임을 확인할 수 있는 멤버십 카드에 한하여 우대가 적용됩니다. \n\n- 플라스틱 또는 모바일 카드와 함께 카드 소유주 확인이 가능한 증빙서류를 지참해 주시기 바랍니다. \n\n- 모바일카드의 경우 APP에서 실행한 카드만 우대가 적용됩니다.\n\n- '롯데월드 어드벤처 부산'은 VIP콕 '롯데월드 어드벤처 부산' 혜택 항목에서 확인하실 수 있습니다.\n\n\n\n대표문의\n\n롯데월드 고객센터 1661-2000"
                  url: "https://adventure.lotteworld.com/kor/main/index.do"
                  tierBenefits:
                    - grade: "VIP콕"
                      context: "본인 종합이용권(1 Day / AFTER 4) 42% 할인\r\n (동반 1인 30% 할인)"
                      isAll: false
                code: BENEFIT_DETAIL_SUCCESS
                status: OK
                message: 혜택 상세 조회에 성공했습니다.
                timestamp: '2025-07-17T10:00:00.000Z'
        '404':
          description: 요청한 리소스를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                benefitNotFound:
                  summary: 혜택이 존재하지 않음
                  value:
                    data: null
                    code: BENEFIT_NOT_FOUND
                    status: NOT_FOUND
                    message: 존재하지 않는 혜택입니다.
                    timestamp: "2025-07-17T18:20:02.2361602"
                storeNotFound:
                  summary: 지점을 찾을 수 없음(storeId가 잘못된 경우)
                  value:
                    data: null
                    code: STORE_NOT_FOUND
                    status: NOT_FOUND
                    message: 지점을 찾을 수 없습니다.
                    timestamp: "2025-07-17T18:20:02.2361602"
                storePartnerMismatch:
                  summary: 지점과 제휴사가 일치하지 않음(partnerId가 잘못된 경우)
                  value:
                    data: null
                    code: STORE_PARTNER_MISMATCH
                    status: BAD_REQUEST
                    message: 지점과 제휴사가 일치하지 않습니다.
                    timestamp: "2025-07-17T18:20:02.2361602"
                benefitOfflineNotFound:
                  summary: 오프라인 사용 가능한 혜택이 없음
                  value:
                    data: null
                    code: BENEFIT_OFFLINE_NOT_FOUND
                    status: NOT_FOUND
                    message: 오프라인 사용 가능한 혜택이 존재하지 않습니다.
                    timestamp: "2025-07-17T18:20:02.2361602"
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: string
          description: "서버에서 정의한 응답 코드"
          example: "SUCCESS_200_1"
        status:
          type: string
          description: "HTTP 상태 메시지"
          example: "OK"
        message:
          type: string
          description: "응답에 대한 설명 메시지"
          example: "요청이 정상적으로 처리되었습니다."
        data:
          type: object
          description: "실제 응답 데이터 (nullable)"
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: "응답 시간"

    # ----------------------------------------------------
    # Request DTOs
    # ----------------------------------------------------
...
    BenefitListResponse:
      type: object
      properties:
        benefitId:
          type: integer
          format: int64
        benefitName:
          type: string
        mainCategory:
          type: string
          enum: [ VIP_COCK, BASIC_BENEFIT ]
        usageType:
          type: string
          enum: [ ONLINE, OFFLINE, BOTH ]
        partnerId:
          type: integer
          format: int64
        category:
          type: string
        image:
          type: string
          format: uri
        tierBenefits:
          type: array
          items:
            $ref: '#/components/schemas/TierBenefitInfo'
        isFavorite:
          type: boolean
        favoriteCount:
          type: integer
          format: int64
    TierBenefitInfo:
      type: object
      properties:
        grade:
          type: string
          enum: [ BASIC, VIP, VVIP, VIP콕 ]
        context:
          type: string
        isAll:
          type: boolean
    PagedResponse_BenefitListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BenefitListResponse'
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalElements:
          type: integer
          format: int64
        hasNext:
          type: boolean
    BenefitDetailResponse:
      type: object
      properties:
        benefitId:
          type: integer
          format: int64
        benefitName:
          type: string
        description:
          type: string
        benefitLimit:
          type: string
        manual:
          type: string
        url:
          type: string
          format: uri
        partnerName:
          type: string
        image:
          type: string
          format: uri
    MapBenefitDetailResponse:
      type: object
      properties:
        benefitId:
          type: integer
          format: int64
        benefitName:
          type: string
        mainCategory:
          type: string
          enum: [ VIP_COCK, BASIC_BENEFIT ]
        manual:
          type: string
        url:
          type: string
          format: uri
        tierBenefits:
          type: array
          items:
            $ref: '#/components/schemas/TierBenefitInfo'
...